fix : 댓글 user 정보, 내 정보(프로필 이미지) #169
