+++ Components/Detail/CommentItem.tsx
@@ -14,7 +14,6 @@ interface CommentType {
   id: string;
   post_id: string;
   user_id: string;
-  user_name: string;
   content: string;
   created_at: string;
 }
@@ -29,6 +28,8 @@ const CommentItem = ({ comment }: CommentItemProps) => {
   const [isEditing, setIsEditing] = useState(false);
   const [editContent, setEditContent] = useState(comment.content);
   const [currentUSERID, setCurrentUSERID] = useState<string | undefined>("");
+  const [userName, setUserName] = useState("");
+  const [userProfileImage, setUserProfileImage] = useState("");
 
   useEffect(() => {
     // 로그인 상태 확인
@@ -41,6 +42,21 @@ const CommentItem = ({ comment }: CommentItemProps) => {
 
     LoginState();
   }, []);
+
+  useEffect(() => {
+    const getUserProfile = async () => {
+      const { data } = await supabase
+        .from("user_profile")
+        .select("*")
+        .eq("user_id", comment.user_id)
+        .single();
+
+      setUserName(data.user_name);
+      setUserProfileImage(data.profile_image);
+    };
+
+    getUserProfile();
+  }, [comment.user_id]);
 
   const { mutate: deleteComment } = useMutation(
     (): any => supabase.from("comment").delete().eq("id", comment.id),
@@ -81,11 +97,15 @@ const CommentItem = ({ comment }: CommentItemProps) => {
 
   return (
     <CommentContainer>
-      <ProfileImage alt="프로필이미지" page="detailPage" />
+      <ProfileImage
+        alt="프로필이미지"
+        page="detailPage"
+        src={userProfileImage}
+      />
       <TextBox>
         <CommentTitle>
           <CommentWrapper>
-            <h3> {comment.user_name} </h3>
+            <h3> {userName} </h3>
             <span> {commentDateView} </span>
           </CommentWrapper>
         </CommentTitle>
