+++ Components/Detail/CommentList.tsx
@@ -6,7 +6,6 @@ interface CommentType {
   id: string;
   post_id: string;
   user_id: string;
-  user_name: string;
   content: string;
   created_at: string;
 }
